+++ src/store/mutations.ts
@@ -10,8 +10,21 @@ import { TreeParent } from '@/types/Tree'
 import { Order } from '@/types/Order'
 
 export default {
+  setLoading (state: ApplicationState, toggle: boolean) {
+    if (toggle) {
+      state.loading += 1
+    } else if (state.loading > 0) {
+      state.loading -= 1
+    }
+  },
   setToast (state: ApplicationState, toast: Toast) {
     state.toast = toast
+  },
+  setProjectNumberRequiredFunction (state: ApplicationState, required: () => boolean) {
+    const projectNumber = state.orderFormFields.find((item) => item.id === 'projectNumber')
+    if (projectNumber) {
+      projectNumber.required = required
+    }
   },
   clearToast (state: ApplicationState) {
     state.toast = null