+++ src/types/ApplicationState.ts
@@ -8,7 +8,7 @@ import { Section } from '@/types/Section.ts'
 import { TreeParent } from '@/types/Tree'
 import FormField from './FormField'
 import { Order } from './Order'
-import { ContextState } from '@molgenis/molgenis-ui-context/src/types'
+import { Context, ContextState } from '@molgenis/molgenis-ui-context/src/types'
 
 export type Toast = {
   type: 'danger' | 'success',
@@ -16,6 +16,7 @@ export type Toast = {
 }
 
 export interface AppState {
+  loading: number,
   order: Order,
   orderFormFields: FormField[]
   variables: { [key:number]: Variable },
@@ -42,6 +43,11 @@ export interface AppState {
   orders: Order[] | null
 }
 
+// we start using the store only after the context is loaded
+interface LoadedContextState extends ContextState {
+    context: Context
+}
+
 export default interface ApplicationState extends AppState {
-  context?: ContextState
+  context: LoadedContextState
 }