+++ tests/unit/store/getters.spec.ts
@@ -1,11 +1,10 @@
 import getters from '@/store/getters'
-import emptyState from '@/store/state'
+import emptyState from '../fixtures/state'
 import Getters from '@/types/Getters'
 import ApplicationState from '@/types/ApplicationState'
 import Variant from '@/types/Variant'
 import Assessment from '@/types/Assessment'
-import { Variable, VariableWithVariants } from '@/types/Variable'
-import { Section } from '@/types/Section'
+import { VariableWithVariants } from '@/types/Variable'
 import { TreeNode } from '@/types/TreeNode'
 
 describe('getters', () => {
@@ -53,9 +52,6 @@ describe('getters', () => {
   }
 
   describe('isSignedIn', () => {
-    it('is false when context is not loaded', () => {
-      expect(getters.isSignedIn({ ...emptyState, context: { context: null } })).toBe(false)
-    })
     it('is false when context is not authenticated', () => {
       expect(getters.isSignedIn({ ...emptyState, context: { context: { authenticated: false } } as any })).toBe(false)
     })